server/requirements.txt
flask
flask_sqlalchemy
flask_cors
protobuf
pymysql
mysql-connector-python
mysqlclient
